Add app directory and move files
