+++ andino_firmware/platformio.ini
@@ -38,8 +38,8 @@ test_build_src = true
 build_src_filter =
   +<drivers/encoder.cpp>
   +<drivers/motor.cpp>
-  +<pid.cpp>
-  +<shell.cpp>
+  +<app/pid.cpp>
+  +<app/shell.cpp>
 
 ; Environment for Arduino Uno.
 [env:uno]
